JIRA-SONIC-11986: [oobe][nous][ec202311.x][nat] code sync of NAT
What I did:
Sync following commit.
(cf0c2d35) JIRA-SONIC-7544: Add a table schema for NatOrch
A new APPL_DB table schema is added to record the update event for NatOrch.

Why I did it:
The table is required for NatOrch in order to fix the issue reported by
the customer.

How I verified it:
The steps for reproduction of the original issue were conducted again to
verify that the issue was resolved. Manual and auto regression tests
were also run on AS5812 for verification.